+++ Runtime/Assist/TableService.cs
@@ -6,15 +6,6 @@ using System.Text.RegularExpressions;
 
 namespace TechTalk.SpecFlow.Assist
 {
-    public interface ITableService
-    {
-        void CompareToSet<T>(Table table, IEnumerable<T> set);
-        void CompareToInstance<T>(Table table, T instance);
-        T CreateInstance<T>(Table table);
-        IEnumerable<T> CreateSet<T>(Table table, Func<T> methodToCreateEachInstance);
-        IEnumerable<T> CreateSet<T>(Table table);
-    }
-
     public class TableService : ITableService
     {
         private readonly Config config;
